moved login back to validationcontroller
